+++ package.json
@@ -98,12 +98,12 @@
     "csurf": "^1.11.0",
     "csv-stringify": "^6.1.0",
     "dayjs": "^1.11.2",
-    "dotenv": "^10.0.0",
+    "dotenv": "^16.0.1",
     "express": "^4.17.3",
     "express-request-id": "^1.4.1",
     "express-session": "^1.17.3",
     "govuk-frontend": "^4.1.0",
-    "helmet": "^4.6.0",
+    "helmet": "^5.1.0",
     "http-errors": "^1.8.1",
     "jquery": "^3.6.0",
     "jwt-decode": "^3.1.2",
@@ -111,7 +111,7 @@
     "nunjucks": "^3.2.3",
     "nunjucks-date-filter": "^0.1.1",
     "openapi-typescript": "^4.5.0",
-    "passport": "^0.4.1",
+    "passport": "^0.6.0",
     "passport-oauth2": "^1.6.1",
     "redis": "^3.1.2",
     "superagent": "^7.1.3"
@@ -130,9 +130,9 @@
     "@types/http-errors": "^1.8.2",
     "@types/jest": "^27.5.1",
     "@types/jsonwebtoken": "^8.5.8",
-    "@types/node": "^14.18.12",
+    "@types/node": "^16.11.36",
     "@types/nunjucks": "^3.2.1",
-    "@types/passport": "^1.0.7",
+    "@types/passport": "^1.0.9",
     "@types/passport-oauth2": "^1.4.11",
     "@types/redis": "^2.8.32",
     "@types/superagent": "^4.1.15",
@@ -140,11 +140,11 @@
     "@typescript-eslint/eslint-plugin": "^5.27.0",
     "@typescript-eslint/parser": "^5.27.0",
     "axe-core": "^4.4.2",
-    "concurrently": "^6.5.1",
-    "cypress": "^8.7.0",
+    "concurrently": "^7.2.1",
+    "cypress": "^9.7.0",
     "cypress-axe": "^0.14.0",
     "cypress-multi-reporters": "^1.6.0",
-    "eslint": "^7.32.0",
+    "eslint": "^8.16.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-config-prettier": "^8.5.0",
     "eslint-import-resolver-typescript": "^2.7.1",
