+++ server/@types/express/index.d.ts
@@ -21,6 +21,8 @@ export declare global {
 
     interface Request {
       verified?: boolean
+      id: string
+      logout(done: (err: unknown) => void): void
     }
   }
 }
