+++ server/authentication/auth.ts
@@ -2,6 +2,7 @@ import passport from 'passport'
 import { Strategy } from 'passport-oauth2'
 import type { RequestHandler } from 'express'
 
+import { inspect } from 'util'
 import config from '../config'
 import generateOauthClientToken from './clientCredentials'
 import type { TokenVerifier } from '../api/tokenVerification'
