+++ server/middleware/setUpAuthentication.ts
@@ -28,15 +28,15 @@ export default function setUpAuth(): Router {
     })(req, res, next)
   )
 
-  const authLogoutUrl = `${config.apis.hmppsAuth.externalUrl}/logout?client_id=${config.apis.hmppsAuth.apiClientId}&redirect_uri=${config.domain}`
-
-  router.use('/logout', (req, res) => {
+  const authUrl = config.apis.hmppsAuth.externalUrl
+  const authSignOutUrl = `${authUrl}/sign-out?client_id=${config.apis.hmppsAuth.apiClientId}&redirect_uri=${config.domain}`
+  router.use('/logout', (req, res, next) => {
     if (req.user) {
-      req.logout()
-      req.session.destroy(() => res.redirect(authLogoutUrl))
-      return
-    }
-    res.redirect(authLogoutUrl)
+      req.logout(err => {
+        if (err) return next(err)
+        return req.session.destroy(() => res.redirect(authSignOutUrl))
+      })
+    } else res.redirect(authSignOutUrl)
   })
 
   router.use((req, res, next) => {
